Moved function definitions to more reasonable headers
