+++ src/rtERenderer/CMakeLists.txt
@@ -2,5 +2,4 @@ target_sources(${PROJECT_NAME} PRIVATE
         rtERenderer.c)
 
 add_subdirectory(creation)
-add_subdirectory(debug)
 add_subdirectory(support)
