+++ src/rtERenderer/creation/CMakeLists.txt
@@ -1,4 +1,2 @@
 target_sources(${PROJECT_NAME} PRIVATE
-        rtER_VK_objectCreation.c
-        rtER_VK_infoCreation.c
-        rtER_VK_readShaderSource.c)
+        rtER_VK_objectCreation.c)
