--- src/rtERenderer/creation/rtER_VK_infoCreation.c
@@ -1,9 +1,0 @@
-#include "rtERenderer/creation/rtER_VK_infoCreation.h"
-
-// If the compiler declines to inline this function, it will look here for the regular definition
-extern inline VkDebugUtilsMessengerCreateInfoEXT rtER_VK_getDebugMessengerCreateInfo(
-        VkDebugUtilsMessageSeverityFlagsEXT severity,
-        VkDebugUtilsMessageTypeFlagsEXT type,
-        PFN_vkDebugUtilsMessengerCallbackEXT callback
-        ); 
-
