--- src/rtERenderer/creation/rtER_VK_readShaderSource.h
@@ -1,7 +1,0 @@
-#ifndef RTER_VK_READSHADERSOURCE_H_
-#define RTER_VK_READSHADERSOURCE_H_
-#include <stdint.h>
-
-unsigned char* rtER_VK_readShaderSource(const char* shaderPath, uint32_t* codeSize);
-
-#endif // RTER_VK_READSHADERSOURCE_H_
