--- src/rtERenderer/debug/CMakeLists.txt
@@ -1,3 +1,0 @@
-target_sources(${PROJECT_NAME} PRIVATE
-        checkValidationLayerSupport.c
-        debugCallback.c)
