+++ src/rtERenderer/support/CMakeLists.txt
@@ -1,3 +1,3 @@
 
 target_sources(${PROJECT_NAME} PRIVATE
-        checkInstanceExtensionSupport.c)
+        rtER_VK_support.c)
