--- src/rtERenderer/support/checkInstanceExtensionSupport.h
@@ -1,7 +1,0 @@
-#ifndef CHECKINSTANCEEXTEXTENSIONSUPPORT_H_
-#define CHECKINSTANCEEXTEXTENSIONSUPPORT_H_
-#include <vulkan/vulkan.h>
-
-VkBool32 rtER_VK_checkInstanceExtensionSupport(const char** requiredInstanceExtensions, uint32_t numRequiredInstanceExtensions);
-
-#endif // CHECKINSTANCEEXTEXTENSIONSUPPORT_H_
